refactor: question 타입 정리
