+++ src/estimations/interfaces/estimation.controller.interface.ts
@@ -1,6 +1,7 @@
-import { Question, QuestionPostDTO } from '#questions/question.types.js';
-import { GetQueries } from '#types/queries.type.js';
 import { EstimationInputDTO } from '#estimations/estimation.types.js';
+import { QuestionPostDTO } from '#questions/types/question.dto.js';
+import { Question } from '#questions/types/question.types.js';
+import { GetQueries } from '#types/queries.type.js';
 
 export interface IEstimationController {
   getQuestions: (id: string, query: GetQueries) => Promise<{ totalCount: number; list: Question[] }>;
