+++ src/swagger/swagger.service.ts
@@ -1,7 +1,7 @@
 import { FilteredDriverOutputDTO } from '#auth/types/filtered.driver.dto.js';
 import { FilteredUserOutputDTO } from '#auth/types/filtered.user.dto.js';
 import { swaggerConfig } from '#configs/swagger.config.js';
-import { QuestionEntity } from '#questions/question.types.js';
+import { QuestionEntity } from '#questions/types/question.types.js';
 import { INestApplication, Injectable } from '@nestjs/common';
 import { OpenAPIObject, SwaggerModule } from '@nestjs/swagger';
 import * as fs from 'fs';
